+++ admin-backend/src/main/java/com/example/demo/dto/CheckValidityResponse.java
@@ -3,11 +3,9 @@ package com.example.demo.dto;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data @Builder
 public class CheckValidityResponse {
-    private BigInteger serialNumber;
+    private Integer serialNumber;
     private Boolean valid;
     private Boolean expired;
     private Boolean started;
